requirements.txt
torch
transformers
diffusers
accelerate
torchvision
numpy
Pillow
tqdm